+++ changelog.md
@@ -1,3 +1,9 @@
+### [0.0.4] - 2022-14-12
+
+#### Added
+- Mapper for Notification repository (database)
+- Mapper for Notification controller (HTTP)
+
 ### [0.0.3] - 2022-12-13
 
 #### Added